rename a folder
