+++ frontend/src/scenes/data-management/properties/EventPropertyDefinitionsTable.scss
